Loading map data from file.

+++ GCU_Programming1_Coursework/MapDefinition.h
@@ -4,7 +4,7 @@
 struct MapDefinition
 {
 	// Defines what tile is placed where.
-	Uint8 tiles[MAP_WIDTH][MAP_HEIGHT];
+	int tiles[MAP_WIDTH][MAP_HEIGHT];
 	// Defines whether a whole tile acts as a collider.
 	bool impassable[MAP_WIDTH][MAP_HEIGHT];
 };
